MMS: Fix The slideshow can't be shown in message
After the MessageItem Pdu Loaded, the MessageItem's AttachmentType
might have changed. So precisely, it's not the same MessageItem
anymore.Since it's not the same MessageItem, we should setImage,
and the slideshow will be shown normally.

After Pdu Loaded, when bindCommonMessage again, the sameItem
argument should be false, because it's not the same MessageItem
anymore.

CRs-Fixed: 741025

Change-Id: Ibef5ca408d86ba76e730fe1fc4117d953d6c74fa